test: 우동 참가자 수 조회 쿼리 udongRepository로 이동(쿼리에서 coalesce하기 위함)

+++ src/main/java/com/hyun/udong/udong/presentation/dto/response/ParticipantCountResponse.java
@@ -1,4 +1,4 @@
 package com.hyun.udong.udong.presentation.dto.response;
 
-public record ParticipantCountResponse(Long udongId, int participantCount) {
+public record ParticipantCountResponse(Long udongId, Long participantCount) {
 }
